backend/requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
astroid==2.14.2
auto-mode==1.0.2
autopep8==2.0.1
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.6
config-parser==0.0.1
configparser==5.3.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
Django==4.2.1
django-cors-headers==3.13.0
django-embed-video==1.4.8
django-environ==0.9.0
django-filter==22.1
django-import-export==3.0.2
django-jazzmin==2.5.0
django-modeltranslation==0.18.7
django-parler==2.3
django-simple-history==3.2.0
django-translated-fields==0.12.0
django-utils-six==2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
drf-nested-routers==0.93.4
drf-writable-nested==0.6.3
drf-yasg==1.20.0
emoji==2.2.0
et-xmlfile==1.1.0
graphene==3.2.1
graphene-django==3.0.0b7
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
install==1.3.5
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
Markdown==3.4.3
MarkupPy==1.14
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.4
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.5.0
pipreqs==0.4.13
platformdirs==3.0.0
promise==2.3
protobuf==3.20.1
psycopg2==2.9.3
pyaes==1.6.1
pyasn1==0.5.0
pycodestyle==2.10.0
PyJWT==2.4.0
pylint==2.16.2
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
PyMySQL==1.0.2
pynvim==0.4.3
pyobjc-core==9.0.1
pyobjc-framework-Cocoa==9.0.1
pyobjc-framework-CoreLocation==9.0.1
pyobjc-framework-Quartz==9.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyUserInput==0.1.11
PyYAML==6.0
ranger-fm==1.9.3
requests==2.28.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
tablib==3.3.0
Telethon==1.28.5
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.10
vk-api==11.9.9
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
yarg==0.1.9
